Added Current c++ rep
Hi there,

I just added the current version of the c++ complete test. A short
description:

*_CONFIG.txt* holds the configuration info for the tests.
*_coreAlgorithms,cpp* holds some basic algorithms used in the test, e.g.
for calculating the ACF or MSE.
*_link_new.cpp* handles the saving of data and logging for windows
machines.
*_test_47.h* simply holds the data that is beeing tested as simple array
(instead of reading it from a csv or equivelent)
*_testProcedure.cpp* holds the test procedure. If the "WriteToLog"
commands are commented out, it should be system independent and only
needs *_coreAlgorithms,cpp* to work.

A complete description follows later. For now the comments in the file
should suffice. The main algorithm called at each step of an ongoing
simulation or in the test-case consecutively in a loop, each time
receiving a new point of data and a new timestamp, is:
"bool OnTheRun_Test (double Data_value, int Data_time)"

-Frederik